Extract string methods from Util.java to StringUtil.java
