+++ src/main/java/skadistats/clarity/io/s2/field/ArrayField.java
@@ -1,9 +1,9 @@
 package skadistats.clarity.io.s2.field;
 
-import skadistats.clarity.io.Util;
 import skadistats.clarity.io.s2.Field;
 import skadistats.clarity.io.s2.FieldType;
 import skadistats.clarity.model.state.ArrayEntityState;
+import skadistats.clarity.util.StringUtil;
 
 public class ArrayField extends Field {
 
@@ -23,12 +23,12 @@ public class ArrayField extends Field {
 
     @Override
     public Integer getChildIndex(String nameSegment) {
-        return Util.stringToArrayIdx(nameSegment);
+        return StringUtil.stringToArrayIdx(nameSegment);
     }
 
     @Override
     public String getChildNameSegment(int idx) {
-        return Util.arrayIdxToString(idx);
+        return StringUtil.arrayIdxToString(idx);
     }
 
     @Override
