+++ src/main/java/skadistats/clarity/io/s2/field/VectorField.java
@@ -1,6 +1,5 @@
 package skadistats.clarity.io.s2.field;
 
-import skadistats.clarity.io.Util;
 import skadistats.clarity.io.decoder.Decoder;
 import skadistats.clarity.io.s2.DecoderHolder;
 import skadistats.clarity.io.s2.DecoderProperties;
@@ -8,6 +7,7 @@ import skadistats.clarity.io.s2.Field;
 import skadistats.clarity.io.s2.FieldType;
 import skadistats.clarity.io.s2.S2DecoderFactory;
 import skadistats.clarity.model.state.ArrayEntityState;
+import skadistats.clarity.util.StringUtil;
 
 public class VectorField extends Field {
 
@@ -37,12 +37,12 @@ public class VectorField extends Field {
 
     @Override
     public Integer getChildIndex(String nameSegment) {
-        return Util.stringToArrayIdx(nameSegment);
+        return StringUtil.stringToArrayIdx(nameSegment);
     }
 
     @Override
     public String getChildNameSegment(int idx) {
-        return Util.arrayIdxToString(idx);
+        return StringUtil.arrayIdxToString(idx);
     }
 
     @Override
